requirements.txt
numpy
matplotlib
seaborn
pandas
pandas_schema
pycountry_convert
progressbar2